+++ include/ignition/math/PID.hh
@@ -30,7 +30,7 @@ namespace ignition
     /// keeps track of PID-error states and control inputs given
     /// the state of a system and a user specified target state.
     /// It includes a user-adjustable command offset term (feed-forward).
-    class IGNITION_VISIBLE PID
+    class IGNITION_MATH_VISIBLE PID
     {
       /// \brief Constructor, zeros out Pid values when created and
       /// initialize Pid-gains and integral term limits:[iMax:iMin]-[I1:I2].