+++ include/ignition/math/SignalStats.hh
@@ -31,7 +31,7 @@ namespace ignition
 
     /// \class SignalStatistic SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Statistical properties of a discrete time scalar signal.
-    class IGNITION_VISIBLE SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalStatistic
     {
       /// \brief Constructor
       public: SignalStatistic();
@@ -78,7 +78,7 @@ namespace ignition
 
     /// \class SignalMaximum SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Computing the maximum value of a discretely sampled signal.
-    class IGNITION_VISIBLE SignalMaximum : public SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalMaximum : public SignalStatistic
     {
       // Documentation inherited.
       public: virtual double Value() const;
@@ -94,7 +94,7 @@ namespace ignition
 
     /// \class SignalMean SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Computing the mean value of a discretely sampled signal.
-    class IGNITION_VISIBLE SignalMean : public SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalMean : public SignalStatistic
     {
       // Documentation inherited.
       public: virtual double Value() const;
@@ -110,7 +110,7 @@ namespace ignition
 
     /// \class SignalMinimum SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Computing the minimum value of a discretely sampled signal.
-    class IGNITION_VISIBLE SignalMinimum : public SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalMinimum : public SignalStatistic
     {
       // Documentation inherited.
       public: virtual double Value() const;
@@ -127,7 +127,7 @@ namespace ignition
     /// \class SignalRootMeanSquare SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Computing the square root of the mean squared value
     /// of a discretely sampled signal.
-    class IGNITION_VISIBLE SignalRootMeanSquare : public SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalRootMeanSquare : public SignalStatistic
     {
       // Documentation inherited.
       public: virtual double Value() const;
@@ -146,7 +146,7 @@ namespace ignition
     /// \brief Computing the maximum of the absolute value
     /// of a discretely sampled signal.
     /// Also known as the maximum norm, infinity norm, or supremum norm.
-    class IGNITION_VISIBLE SignalMaxAbsoluteValue : public SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalMaxAbsoluteValue : public SignalStatistic
     {
       // Documentation inherited.
       public: virtual double Value() const;
@@ -163,7 +163,7 @@ namespace ignition
     /// \class SignalVariance SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Computing the incremental variance
     /// of a discretely sampled signal.
-    class IGNITION_VISIBLE SignalVariance : public SignalStatistic
+    class IGNITION_MATH_VISIBLE SignalVariance : public SignalStatistic
     {
       // Documentation inherited.
       public: virtual double Value() const;
@@ -182,7 +182,7 @@ namespace ignition
 
     /// \class SignalStats SignalStats.hh ignition/math/SignalStats.hh
     /// \brief Collection of statistics for a scalar signal.
-    class IGNITION_VISIBLE SignalStats
+    class IGNITION_MATH_VISIBLE SignalStats
     {
       /// \brief Constructor
       public: SignalStats();