+++ src/graph/GraphDirected_TEST.cc
@@ -50,30 +50,30 @@ TEST(GraphTest, Edges)
     {
       case 0:
       {
-        EXPECT_EQ(0, edge.Tail());
-        EXPECT_EQ(0, edge.Head());
-        EXPECT_EQ(1.0, edge.Data());
+        EXPECT_EQ(0u, edge.Tail());
+        EXPECT_EQ(0u, edge.Head());
+        EXPECT_DOUBLE_EQ(1.0, edge.Data());
         break;
       }
       case 1:
       {
-        EXPECT_EQ(0, edge.Tail());
-        EXPECT_EQ(1, edge.Head());
-        EXPECT_EQ(2.0, edge.Data());
+        EXPECT_EQ(0u, edge.Tail());
+        EXPECT_EQ(1u, edge.Head());
+        EXPECT_DOUBLE_EQ(2.0, edge.Data());
         break;
       }
       case 2:
       {
-        EXPECT_EQ(1, edge.Tail());
-        EXPECT_EQ(2, edge.Head());
-        EXPECT_EQ(3.0, edge.Data());
+        EXPECT_EQ(1u, edge.Tail());
+        EXPECT_EQ(2u, edge.Head());
+        EXPECT_DOUBLE_EQ(3.0, edge.Data());
         break;
       }
       case 3:
       {
-        EXPECT_EQ(2, edge.Tail());
-        EXPECT_EQ(0, edge.Head());
-        EXPECT_EQ(4.0, edge.Data());
+        EXPECT_EQ(2u, edge.Tail());
+        EXPECT_EQ(0u, edge.Head());
+        EXPECT_DOUBLE_EQ(4.0, edge.Data());
         break;
       }
       default:
@@ -207,17 +207,17 @@ TEST(GraphTest, IncidentsFrom)
       case 2:
       {
         auto vertices = edge.Vertices();
-        EXPECT_EQ(1, vertices.first);
-        EXPECT_EQ(0, vertices.second);
-        EXPECT_EQ(3.0, edge.Data());
+        EXPECT_EQ(1u, vertices.first);
+        EXPECT_EQ(0u, vertices.second);
+        EXPECT_DOUBLE_EQ(3.0, edge.Data());
         break;
       }
       case 3:
       {
         auto vertices = edge.Vertices();
-        EXPECT_EQ(1, vertices.first);
-        EXPECT_EQ(2, vertices.second);
-        EXPECT_EQ(4.0, edge.Data());
+        EXPECT_EQ(1u, vertices.first);
+        EXPECT_EQ(2u, vertices.second);
+        EXPECT_DOUBLE_EQ(4.0, edge.Data());
         break;
       }
       default:
@@ -263,17 +263,17 @@ TEST(GraphTest, IncidentsTo)
       case 0:
       {
         auto vertices = edge.Vertices();
-        EXPECT_EQ(0, vertices.first);
-        EXPECT_EQ(0, vertices.second);
-        EXPECT_EQ(1.0, edge.Data());
+        EXPECT_EQ(0u, vertices.first);
+        EXPECT_EQ(0u, vertices.second);
+        EXPECT_DOUBLE_EQ(1.0, edge.Data());
         break;
       }
       case 3:
       {
         auto vertices = edge.Vertices();
-        EXPECT_EQ(2, vertices.first);
-        EXPECT_EQ(0, vertices.second);
-        EXPECT_EQ(4.0, edge.Data());
+        EXPECT_EQ(2u, vertices.first);
+        EXPECT_EQ(0u, vertices.second);
+        EXPECT_DOUBLE_EQ(4.0, edge.Data());
         break;
       }
       default:
@@ -334,9 +334,9 @@ TEST(GraphTest, AddEdge)
   auto &e2 = graph.AddEdge({2, 0}, 4.0);
 
   // Check the edge content.
-  EXPECT_EQ(2.0, e0.Data());
-  EXPECT_EQ(3.0, e1.Data());
-  EXPECT_EQ(4.0, e2.Data());
+  EXPECT_DOUBLE_EQ(2.0, e0.Data());
+  EXPECT_DOUBLE_EQ(3.0, e1.Data());
+  EXPECT_DOUBLE_EQ(4.0, e2.Data());
 
   // Change some content and verity it.
   e2.Data() = 5.0;
@@ -346,8 +346,8 @@ TEST(GraphTest, AddEdge)
   EXPECT_DOUBLE_EQ(6.0, edge.Weight());
 
   // Check that the edges point to the right vertices.
-  EXPECT_EQ(0, e0.Tail());
-  EXPECT_EQ(1, e0.Head());
+  EXPECT_EQ(0u, e0.Tail());
+  EXPECT_EQ(1u, e0.Head());
 
   auto edges = graph.Edges();
   EXPECT_EQ(3u, edges.size());
@@ -447,7 +447,7 @@ TEST(GraphTest, RemoveVertices)
   });
 
   // Try to remove a node with a name that doesn't exist.
-  EXPECT_EQ(graph.RemoveVertices("wrong_name"), 0);
+  EXPECT_EQ(graph.RemoveVertices("wrong_name"), 0u);
   EXPECT_EQ(4u, graph.Vertices().size());
   EXPECT_EQ(1u, graph.AdjacentsFrom(1).size());